@@ -29,7 +29,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -40,14 +40,14 @@ class MyHomePage extends StatefulWidget {
   // used by the build method of the State. Fields in a Widget subclass are
   // always marked "final".
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  PackageInfo packageInfo;
+  PackageInfo? packageInfo;
 
   @override
   void initState() {
@@ -71,15 +71,15 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         // Here we take the value from the MyHomePage object that was created by
         // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
+        title: Text(widget.title!),
       ),
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: <Widget>[
           if (packageInfo != null) ...[
-            _buildRow('App name', packageInfo.appName),
-            _buildRow('Build Number', packageInfo.buildNumber),
-            _buildRow('Version', packageInfo.version),
+            _buildRow('App name', packageInfo!.appName),
+            _buildRow('Build Number', packageInfo!.buildNumber),
+            _buildRow('Version', packageInfo!.version),
           ],
         ],
       ),
@@ -99,7 +99,7 @@ class _MyHomePageState extends State<MyHomePage> {
           ),
           Text(
             value,
-            style: Theme.of(context).textTheme.subtitle1.copyWith(
+            style: Theme.of(context).textTheme.subtitle1!.copyWith(
                   color: Colors.red,
                 ),
           ),
